Rename transaction::components to transaction::impls
