+++ crates/relayer/relayer-components/src/transaction/mod.rs
@@ -2,5 +2,5 @@
    Constructs for the transaction context.
 */
 
-pub mod components;
+pub mod impls;
 pub mod traits;
